+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- BREAKING CHANGE: AdcsCertificateAuthoritySettings resource requires WMF 5.0
+  so ActiveDirectoryCSDsc module minimum requirements WMF 5.0.
+- Added new resource AdcsCertificateAuthoritySettings - see
+  [Issue #13](https://github.com/PowerShell/ActiveDirectoryCSDsc/issues/13).
 - Added new resource AdcsTemplate.
 - Replaced `switch` blocks with `if` blocks for evaluating 'Ensure' parameter
   because switch was missing `break` - fixes [Issue #87](https://github.com/PowerShell/ActiveDirectoryCSDsc/issues/87).
@@ -15,8 +19,6 @@
 - AdcsCertificationAuthority:
   - Integration tests updated to create test user account in administrators
     group to make test execution easier.
-- Added new resource AdcsCertificateAuthoritySettings - see
-  [Issue #13](https://github.com/PowerShell/ActiveDirectoryCSDsc/issues/13).
 
 ## 3.3.0.0